Nit refactor and clean up
