+++ apps/studio/components/interfaces/Functions/EdgeFunctionsListItem.utils.ts
@@ -12,5 +12,9 @@ export function formatErrorRate(value: number): string {
   if (value === 0) return '0%'
   if (value >= 100) return '100%'
   if (value < 0.1) return '<0.1%'
-  return `${value.toFixed(1)}%`
+
+  const rounded = Number(value.toFixed(1))
+  const clamped = Math.min(rounded, 100)
+  if (clamped >= 100) return '100%'
+  return `${clamped.toFixed(1)}%`
 }
